# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/subscriptions": {
      "post": {
        "operationId": "SubscriptionsController_create",
        "summary": "Create a new subscription",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The subscription has been successfully created."
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "tags": [
          "subscriptions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "SubscriptionsController_findAll",
        "summary": "Get all subscriptions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all subscriptions."
          },
          "404": {
            "description": "Subscriptions not found."
          }
        },
        "tags": [
          "subscriptions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/subscriptions/{id}": {
      "get": {
        "operationId": "SubscriptionsController_findOne",
        "summary": "Get a subscription by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return a subscription."
          },
          "404": {
            "description": "Subscription not found."
          }
        },
        "tags": [
          "subscriptions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "SubscriptionsController_update",
        "summary": "Update a subscription by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The subscription has been successfully updated."
          },
          "404": {
            "description": "Subscription not found."
          }
        },
        "tags": [
          "subscriptions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "SubscriptionsController_delete",
        "summary": "Delete a subscription by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The subscription has been successfully deleted."
          },
          "404": {
            "description": "Subscription not found."
          }
        },
        "tags": [
          "subscriptions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/subscriptions/me": {
      "get": {
        "operationId": "SubscriptionsController_findMySubscription",
        "summary": "Get the subscription of the current user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return the subscription of the current user."
          },
          "404": {
            "description": "Subscription not found."
          }
        },
        "tags": [
          "subscriptions"
        ]
      }
    },
    "/subscriptions/cancel-subscription/{subscriptionId}": {
      "delete": {
        "operationId": "SubscriptionsController_cancelSubscription",
        "summary": "Cancel a subscription by id",
        "parameters": [
          {
            "name": "subscriptionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The subscription has been successfully canceled."
          },
          "404": {
            "description": "Subscription not found."
          }
        },
        "tags": [
          "subscriptions"
        ]
      }
    }
  },
  "info": {
    "title": "Api ZenFlow",
    "description": "Api for a CRM/Forum application with modules.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateSubscriptionDto": {
        "type": "object",
        "properties": {
          "customer": {
            "type": "string",
            "description": "The customer ID for the subscription"
          },
          "plan": {
            "type": "string",
            "description": "The plan ID for the subscription"
          },
          "stripeCustomerId": {
            "type": "string",
            "description": "The stripe customer id"
          },
          "stripeSubscriptionId": {
            "type": "string",
            "description": "The stripe subscription id"
          },
          "startDate": {
            "type": "string",
            "description": "The start date of the subscription"
          },
          "endDate": {
            "type": "string",
            "description": "The end date of the subscription"
          }
        },
        "required": [
          "customer",
          "plan"
        ]
      },
      "UpdateSubscriptionDto": {
        "type": "object",
        "properties": {
          "customer": {
            "type": "string",
            "description": "The customer ID for the subscription"
          },
          "plan": {
            "type": "string",
            "description": "The plan ID for the subscription"
          },
          "stripeCustomerId": {
            "type": "string",
            "description": "The stripe customer id"
          },
          "stripeSubscriptionId": {
            "type": "string",
            "description": "The stripe subscription id"
          },
          "startDate": {
            "type": "string",
            "description": "The start date of the subscription"
          },
          "endDate": {
            "type": "string",
            "description": "The end date of the subscription"
          }
        }
      }
    }
  }
}